Moved keywords and sentences files into lib folder.
