+++ src/rake.js
@@ -1,15 +1,16 @@
 import defaultOptions from './options'
-import splitSentences from './sentences'
-import findCandidateKeywords from './keywords'
+import splitSentences from './lib/sentences'
+import findCandidateKeywords from './lib/keywords'
+import { calculateWordScores } from './lib/scores'
 
 const rake = (text, stopWords = [], overrideOptions = {}) => {
     const options = Object.assign(defaultOptions, overrideOptions)
 
     const sentences = splitSentences(text)
 
-    const candidateKeywords = findCandidateKeywords(sentences, stopWords, options)
+    const candidatePhrases = findCandidateKeywords(sentences, stopWords, options)
 
-    const wordScores = calculateWordScores(candidateKeywords)
+    const wordScores = calculateWordScores(candidatePhrases)
 }
 
 export default rake
